requirements.txt
pandas>=1.2.4
numpy>=1.20.3
matplotlib>=3.4.2
psycopg2-binary>=2.8.6